% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runttest.R
\name{.runttest}
\alias{.runttest}
\title{Compute independent-samples t-test}
\usage{
.runttest(y, group)
}
\arguments{
\item{y}{Vector of y values}

\item{group}{Vector indicating group membership as 0 or 1}
}
\description{
Computes p-value from two-sided t-test from sample
}
