% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectpvalue.R
\name{.selectpvalue}
\alias{.selectpvalue}
\title{Select smallest, smallest significant, and first significant p-value from a vector of p-hacked p-values}
\usage{
.selectpvalue(ps, alpha)
}
\arguments{
\item{ps}{Vector of p values}

\item{alpha}{Significance level (default: 0.05)}
}
\description{
Takes a vector of p-values and selects the smallest, first significant, or smallest significant p-value.
}
