% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pcurves.R
\name{plot_pcurves}
\alias{plot_pcurves}
\title{Plot original and fitted p-curves}
\usage{
plot_pcurves(simdat, poriginal, n_best = 10)
}
\arguments{
\item{simdat}{Result of simulation function}

\item{poriginal}{Original p-curve (vector)}
}
\description{
Create a plot displaying the original and a number of fitted p-curves
}
