% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Compute RMSE between two p-curves}
\usage{
rmse(v1, v2)
}
\arguments{
\item{v1}{First vector (original)}

\item{v2}{Second vector (fitted values)}
}
\description{
Calculates square root of mean of squared differences between two vectors
}
