% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_cached_file}
\alias{save_cached_file}
\title{Save an R Object to a Sharded Cache File}
\usage{
save_cached_file(object, file, compress = "xz")
}
\arguments{
\item{object}{The R object to be serialized and saved.}

\item{file}{A string specifying the file path (including filename and
extension) where the object will be stored. Typically an \code{.rds} file.}

\item{compress}{A single string indicating the compression algorithm to
use. Passed to the \code{compress} argument of \code{\link[=saveRDS]{saveRDS()}}. Common options
include \code{"none"}, \code{"gzip"}, \code{"bzip2"}, and \code{"xz"}. Defaults to \code{"xz"}.}
}
\description{
Wrapper around \link[base:readRDS]{base::saveRDS} that writes an R object to a specified file
path, using a chosen compression algorithm. Intended for use with
cache filenames generated by \code{\link[=get_cache_filename]{get_cache_filename()}}.
}
