% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingDV.R
\name{sim.multDVhack}
\alias{sim.multDVhack}
\title{Simulate p-Hacking with multiple dependent variables}
\usage{
sim.multDVhack(
  nvar,
  r,
  d,
  strategy = "firstsig",
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{nvar}{Number of dependent variables (columns) in the data frame}

\item{r}{Desired correlation between the dependent variables (scalar)}

\item{d}{Desired population effect size (standardized mean difference between grouping variable levels)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{iter}{Number of simulation iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Outputs a vector containing the p-hacked p-values
}
